@@ -1,11 +1,11 @@
 {
   "name": "webpack",
-  "version": "1.0.0",
+  "version": "5.73.0",
   "description": "A dope webpack boilerplate",
   "main": "index.js",
   "scripts": {
-    "watch": "rimraf dist && webpack --watch",
-    "start": "webpack-dev-server  --open --color ",
+    "watch": "rimraf dist && webpack serve --watch --progress",
+    "start": "webpack-dev-server  --open --color",
     "build": "rimraf dist &&  webpack --progress --profile",
     "server": "nodemon server/server.js",
     "server-prod": "node server/server.js"
@@ -17,32 +17,23 @@
   "author": "James Dempsey",
   "license": "UNLICENSED",
   "dependencies": {
-    "@ionic/react": "^5.5.4",
-    "@ionic/react-router": "^5.5.4",
-    "axios": "^0.21.1",
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "axios": "^0.27.2",
+    "body-parser": "^1.20.0",
+    "express": "^4.18.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-loadable": "^5.5.0",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "^6.3.0"
   },
   "babel": {
     "presets": [
+      "@babel/env",
       [
-        "env",
+        "@babel/react",
         {
-          "targets": {
-            "browsers": [
-              ">1%",
-              "last 3 versions"
-            ]
-          }
+          "runtime": "automatic"
         }
-      ],
-      "stage-2",
-      "latest",
-      "react"
+      ]
     ],
     "plugins": [
       "syntax-dynamic-import",
@@ -50,25 +41,28 @@
     ]
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.0.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
+    "babel-loader": "^8.2.5",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.6.1",
     "babel-preset-latest": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
-    "css-loader": "^0.28.11",
-    "file-loader": "^1.1.11",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.4.0",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.21.0",
-    "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack": "^4.29.3",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.1.4"
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "rimraf": "^3.0.2",
+    "style-loader": "^3.3.1",
+    "webpack": "5.73.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.2"
+  },
+  "engines": {
+    "node": ">= 18.2.0",
+    "npm": ">= 8.9.0"
   }
 }
